split 2 classes into 2 files
